Fix the issue where theme toggle had to use different container tags for sun and moon.
I _think_ this is now working due to the fact that the inline script runs at runtime on user's browser, rather than during Gatsby's
"build" phase, and it sets the initial theme which is then picked up by the theme hook. This _feels_ like the correct behaviour, but
still not entirely sure what the deal was with having to need different container tags for sun and moon in the theme toggle!